Create components/IconButton/CopyIconButton
